opencue/jobs/requirements.txt
# OpenCue Python 依赖
# 不指定版本，使用最新版
opencue-pycue
opencue-pyoutline
